+++ resources/views/components/layouts/live.blade.php
@@ -5,6 +5,7 @@
 
 <head>
     <meta charset="utf-8">
+    <x-theme-init />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="csrf-token" content="{{ csrf_token() }}">
     <title>{{ $title ? $title . ' - ' . config('app.name') : config('app.name') }}</title>